feat:给dao层添加count方法，memberService增加返回默认数据,添加PageVO对象

+++ src/main/java/com/qgailab/service/impl/MemberServiceImpl.java
@@ -1,22 +1,25 @@
 package com.qgailab.service.impl;
 
 import com.github.pagehelper.PageHelper;
-import com.qgailab.dao.LeaderMapper;
+import com.qgailab.dao.ImageMapper;
 import com.qgailab.dao.MemberMapper;
 import com.qgailab.model.dto.ServiceResult;
+import com.qgailab.model.po.Image;
 import com.qgailab.model.po.Member;
-import com.qgailab.model.po.Moment;
+import com.qgailab.model.po.PageVO;
 import com.qgailab.service.MemberService;
 import com.qgailab.service.constants.Message;
 import com.qgailab.util.UUIDUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
+import java.util.LinkedList;
 import java.util.List;
 
 /**
- * @description 成员增删查改
  * @author < a href=" ">郭沛</ a>
+ * @description 成员增删查改
  * @date 2019-07-26 22:23
  */
 @Service
@@ -24,6 +27,9 @@ public class MemberServiceImpl implements MemberService {
 
     @Autowired
     private MemberMapper memberMapper;
+
+    @Autowired
+    private ImageMapper imageMapper;
 
     /**
      * 负责插入一条成员信息
@@ -44,10 +50,18 @@ public class MemberServiceImpl implements MemberService {
                 return new ServiceResult(401, Message.name_not_null);
             }
             member.setUuid(UUIDUtils.getUUID());
+            //没有图片的插入默认图片
+            if (member.getImages() == null || member.getImages().size() == 0) {
+                Image image = new Image();
+                image.setFilename(imagePath);
+                image.setFilename(member.getUuid());
+                imageMapper.insertSelective(image);
+            }
+
             if (memberMapper.insertSelective(member) != 1) {
                 return new ServiceResult(402, Message.database_exception);
             }
-        }catch (Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             return new ServiceResult(500, Message.please_retry);
         }
@@ -66,15 +80,19 @@ public class MemberServiceImpl implements MemberService {
      */
     @Override
     public ServiceResult removeMember(Long id) {
-        Member member = memberMapper.selectByPrimaryKey(id);
+        if (id == null) {
+            return new ServiceResult(400, Message.parameter_not_enough);
+        }
+        Member member;
         try {
+            member = memberMapper.selectByPrimaryKey(id);
             if (member == null) {
                 return new ServiceResult(401, Message.member_not_found);
             }
             if (memberMapper.deleteByPrimaryKey(id) != 1) {
                 return new ServiceResult(402, Message.database_exception);
             }
-        }catch (Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             return new ServiceResult(500, Message.please_retry);
         }
@@ -97,7 +115,7 @@ public class MemberServiceImpl implements MemberService {
             if (member == null) {
                 return new ServiceResult(401, Message.member_not_found);
             }
-        }catch (Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             return new ServiceResult(500, Message.please_retry);
         }
@@ -125,16 +143,20 @@ public class MemberServiceImpl implements MemberService {
             if (memberMapper.updateByPrimaryKeySelective(member) != 1) {
                 return new ServiceResult(402, Message.database_exception);
             }
-        }catch (Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             return new ServiceResult(500, Message.please_retry);
         }
         return new ServiceResult(200, Message.success, member);
     }
 
+    @Value(value = "#{conf.image}")
+    private String imagePath;
+
     /**
      * 负责查询多个成员信息
-     * @param page  页数
+     *
+     * @param page     页数
      * @param pageSize 一页最大记录数
      * @return: ServiceResult
      * @Author: gp
@@ -142,18 +164,35 @@ public class MemberServiceImpl implements MemberService {
      */
     @Override
     public ServiceResult listMember(int page, int pageSize) {
-        if (page < 0 ){
+        if (page <= 0) {
             return new ServiceResult(400, Message.page_invalid);
         }
+        if (pageSize <= 0) {
+            return new ServiceResult(401, Message.pageSize_invalid);
+        }
         List<Member> memberList;
+        int count;
         try {
             PageHelper.startPage(page, pageSize);
             memberList = memberMapper.listPage(page * pageSize, pageSize);
-        }catch (Exception e) {
+            count = memberMapper.selectCount();
+            //创建一个默认的成员
+            Member member = new Member();
+            member.setField("");
+            member.setName("");
+            member.setGrade("");
+            Image image = new Image();
+            image.setFilename(imagePath);
+            List<Image> imageList = new LinkedList<>();
+            imageList.add(image);
+            member.setImages(imageList);
+            memberList.add(member);
+
+        } catch (Exception e) {
             e.printStackTrace();
             return new ServiceResult(500, Message.please_retry);
         }
-        return new ServiceResult(200, Message.success, memberList);
+        return new ServiceResult(200, Message.success, new PageVO(count / pageSize, memberList));
     }
